A bunch of preparations; introduced GrothendieckTopos, moved classifier there.
